Feat: ajuste nas rotas, controler, services, client e  schema status divergencia
