+++ src/ConferenciaCega/client/index.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import 'dotenv/config';
-const url = process.env.API_URL;
-
+//const url = process.env.API_URL;
+const url = 'http://164.152.245.77:8000/quality/concentrador_node';
 
 export class OTClient {
     constructor(baseURL) {
@@ -244,6 +244,36 @@ export class OTClient {
             IDUSRAJUSTE,
             IDSTATUSOT,
             IDRESUMOOT
+        });
+        return response.data;
+    }
+
+    async statusDivergenciaClient(
+        DESCRICAODIVERGENCIA,
+        IDUSRCRIACAO,
+        STATIVO,
+
+    ) {
+
+        const response = await this.api.post(`api/conferencia-cega/status-divergencia.xsjs`, {
+            DESCRICAODIVERGENCIA,
+            IDUSRCRIACAO,
+            STATIVO,
+        });
+        return response.data;
+    }
+
+    async statusEditarDivergenciaClient(
+        DESCRICAODIVERGENCIA,
+        IDSTATUSDIVERGENCIA,
+        STATIVO,
+
+    ) {
+
+        const response = await this.api.put(`/api/conferencia-cega/status-divergencia.xsjs`, {
+            DESCRICAODIVERGENCIA,
+            IDSTATUSDIVERGENCIA,
+            STATIVO,
         });
         return response.data;
     }
